refactor(web): move settings-error util to shared/
Used by 8 settings tab files plus the soon-to-be-shared MembersTab;
shared/ is the correct home for cross-route utilities.
