+++ apps/web/app/(dashboard)/settings/_components/members-tab.tsx
@@ -38,7 +38,7 @@ import {
 import { toast } from "sonner";
 import { useAuthStore } from "@/features/auth";
 import { useWorkspaceManagement } from "@/features/workspace";
-import { getSettingsErrorMessage } from "./settings-error";
+import { getSettingsErrorMessage } from "@/shared/settings-error";
 
 const roleConfig: Record<MemberRole, { label: string; icon: typeof Crown; description: string }> = {
   owner: { label: "所有者", icon: Crown, description: "完全访问权限，管理所有设置" },
